backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.52.0
python-dotenv==1.0.1
python-pptx==1.0.2
python-docx==1.1.2
pydantic==2.10.4
httpx==0.28.1
